VidSynth/push_service/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
pydantic